Refactor and dev tools. Added helper functions for api operations. Added utils for dropping tables. Updated table definitions.

+++ src/server/model/task-model.ts
@@ -1,20 +1,25 @@
 import mongoose, { Schema, Document as IDocument } from 'mongoose';
 
+import { dropAllTables, dropTaskTable } from '../utils/dev-utils';
+
 interface TaskSchemaModel extends Task, IDocument {
 }
 
 const TaskSchema = new Schema<TaskSchemaModel>({
-    frequency: { type: String, required: true },
     householdID: { type: String, required: true },
+    frequency: { type: String, required: true },
     points: { type: Number, required: true },
-    title: { type: String, required: true },
+    taskName: { type: String, required: true },
     desc: { type: String },
 });
 
-TaskSchema.index({ householdID: 1, title: 1 }, { unique: true });
+TaskSchema.index({ householdID: 1, taskName: 1 }, { unique: true });
 
 const TaskModel = mongoose.model<TaskSchemaModel>('task', TaskSchema);
 
-TaskModel.collection.drop();
+if (dropAllTables || dropTaskTable) {
+    TaskModel.collection.drop();
+}
+
 
 export default TaskModel;
